requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-doctr[viz] @ git+https://github.com/mindee/doctr.git
matplotlib
pillow>=9.0.0
numpy>=1.21.0
torch>=1.9.0
python-multipart>=0.0.5
pypdf>=4.0.0
ocrmypdf>=15.0.0
openai>=1.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0